backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0
asyncpg>=0.30.0
alembic>=1.14.0
redis>=5.0
python-multipart>=0.0.18
httpx>=0.28.0
langchain>=0.3.0
langfuse>=2.0
pgvector>=0.3.0
python-dotenv>=1.0
psycopg2-binary>=2.9
google-genai>=1.0
openai>=1.0
pyyaml>=6.0
python-docx>=1.0
